All the silver parts of monto.
I'll squash this later to avoid the commit noise of the monto2
directory.
