+++ grammars/silver/support/monto/products/Product.sv
@@ -5,42 +5,58 @@ import lib:json;
 synthesized attribute productName :: String;
 nonterminal ProductValue with json, productName;
 
+synthesized attribute productLang :: String;
+synthesized attribute productPath :: String;
 synthesized attribute productValue :: ProductValue;
-nonterminal Product with json, productValue;
+nonterminal Product with json, productName, productLang, productPath, productValue;
 
 abstract production product
-top::Product ::= name::String value::ProductValue metas::[MetaItem] language::String path::String
+top::Product ::= value::ProductValue metas::[MetaItem] language::String path::String
 {
+  top.productName = top.productValue.productName;
+  top.productLang = language;
+  top.productPath = path;
+  top.productValue = value;
   top.json = jsonObject(
     [ pair("contents", value.json)
-    , pair("language", jsonString(language))
-    , pair("name", jsonString("name"))
+    , pair("language", jsonString(top.productLang))
+    , pair("name", jsonString(top.productName))
     , pair("meta", jsonArray(map((.json), metas)))
-    , pair("path", jsonString(path))
+    , pair("path", jsonString(top.productPath))
     ]);
-  top.productValue = value;
 }
 
-nonterminal ProductDescriptor with json;
+nonterminal ProductDescriptor with json, productName, productLang;
 
 abstract production productDescriptor
 top::ProductDescriptor ::= name::String language::String
 {
+  top.productName = name;
+  top.productLang = language;
   top.json = jsonObject(
-    [ pair("name", jsonString("name"))
-    , pair("language", jsonString("language"))
+    [ pair("name", jsonString(top.productName))
+    , pair("language", jsonString(top.productLang))
     ]);
 }
 
-nonterminal ProductIdentifier with json;
+function productDescriptorEq
+Boolean ::= l::ProductDescriptor r::ProductDescriptor
+{
+  return l.productName == r.productName && l.productLang == r.productLang;
+}
+
+nonterminal ProductIdentifier with json, productName, productLang, productPath;
 
 abstract production productIdentifier
 top::ProductIdentifier ::= name::String language::String path::String
 {
+  top.productName = name;
+  top.productLang = language;
+  top.productPath = path;
   top.json = jsonObject(
-    [ pair("name", jsonString("name"))
-    , pair("language", jsonString("language"))
-    , pair("path", jsonString("path"))
+    [ pair("name", jsonString(top.productName))
+    , pair("language", jsonString(top.productLang))
+    , pair("path", jsonString(top.productPath))
     ]);
 }
