+++ runtime/java/src/monto2/RegistrationFailedException.java
